# apps/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/classrooms": {
      "post": {
        "operationId": "ClassroomsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Аудитория успешно создана"
          },
          "400": {
            "description": "Некорректные данные"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новую аудиторию",
        "tags": [
          "Classrooms"
        ]
      },
      "get": {
        "operationId": "ClassroomsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список всех аудиторий"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все аудитории",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/building/{building}": {
      "get": {
        "operationId": "ClassroomsController_findByBuilding",
        "parameters": [
          {
            "name": "building",
            "required": true,
            "in": "path",
            "description": "Название здания",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Аудитории в указанном здании"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить аудитории по зданию",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/type/{type}": {
      "get": {
        "operationId": "ClassroomsController_findByType",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "description": "Тип аудитории",
            "schema": {
              "enum": [
                "LECTURE",
                "PRACTICE",
                "COMPUTER",
                "LABORATORY",
                "OTHER"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аудитории указанного типа"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить аудитории по типу",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/capacity/{minCapacity}": {
      "get": {
        "operationId": "ClassroomsController_findByCapacity",
        "parameters": [
          {
            "name": "maxCapacity",
            "required": false,
            "in": "query",
            "description": "Максимальная вместимость",
            "schema": {}
          },
          {
            "name": "minCapacity",
            "required": true,
            "in": "path",
            "description": "Минимальная вместимость",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Аудитории с подходящей вместимостью"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Найти аудитории по вместимости",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/available/{dayOfWeek}/{startTime}/{endTime}": {
      "get": {
        "operationId": "ClassroomsController_findAvailableClassrooms",
        "parameters": [
          {
            "name": "endTime",
            "required": true,
            "in": "path",
            "description": "Время окончания (HH:MM)",
            "schema": {}
          },
          {
            "name": "startTime",
            "required": true,
            "in": "path",
            "description": "Время начала (HH:MM)",
            "schema": {}
          },
          {
            "name": "dayOfWeek",
            "required": true,
            "in": "path",
            "description": "День недели (1-7)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Список свободных аудиторий"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Найти свободные аудитории на указанное время",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/by-equipment": {
      "post": {
        "operationId": "ClassroomsController_findByEquipment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Аудитории с указанным оборудованием"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Найти аудитории по оборудованию",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/statistics": {
      "get": {
        "operationId": "ClassroomsController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Статистика аудиторий"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить статистику по аудиториям",
        "tags": [
          "Classrooms"
        ]
      }
    },
    "/classrooms/{id}": {
      "get": {
        "operationId": "ClassroomsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID аудитории",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Данные аудитории с расписанием"
          },
          "404": {
            "description": "Аудитория не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить аудиторию по ID",
        "tags": [
          "Classrooms"
        ]
      },
      "patch": {
        "operationId": "ClassroomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID аудитории",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Аудитория успешно обновлена"
          },
          "404": {
            "description": "Аудитория не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить аудиторию",
        "tags": [
          "Classrooms"
        ]
      },
      "delete": {
        "operationId": "ClassroomsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID аудитории",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Аудитория успешно удалена"
          },
          "404": {
            "description": "Аудитория не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить аудиторию",
        "tags": [
          "Classrooms"
        ]
      }
    },
...
    "/schedule/classroom/{classroomId}": {
      "get": {
        "operationId": "ScheduleController_findByClassroom",
        "parameters": [
          {
            "name": "classroomId",
            "required": true,
            "in": "path",
            "description": "ID аудитории",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Расписание аудитории"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить расписание аудитории",
        "tags": [
          "Schedule"
        ]
      }
    },
...
    "/ai-assistant/session": {
      "post": {
        "operationId": "AiAssistantController_createSession",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Ephemeral token создан успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "client_secret": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": "string",
                          "description": "Ephemeral token"
                        },
                        "expires_at": {
                          "type": "number",
                          "description": "Время истечения токена (Unix timestamp)"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать ephemeral token для OpenAI Realtime API",
        "tags": [
          "AI Assistant"
        ]
      }
    }
  },
  "info": {
    "title": "Multi-Tenant ABAI API",
...
    "version": "1.0.0",
    "contact": {
      "name": "ABAI Support Team",
      "url": "https://abai.edu.kz",
      "email": "support@abai.edu.kz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.abai.edu.kz",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен",
        "in": "header"
      }
    },
...
  }
}